feat: allow disabling deletion after install on some platforms
Mac:

 If deleteAfterInstall is set to false, the downloaded file
 will not be deleted after installing it. By default it's true.

Linux:

 Delete after install argument is ignored on Linux, since the downloaded
 AppImage always replaces the existing one.  It's faster than copying
 and keeping another copy.  Successful replacement should not result in
 update failure, while unsuccessful replacement will keep the file in
 the original location.

Windows:

 On Windows, deleteAfterInstall is ignored, since we launch the
 installer and quit, thus we cannot delete the file.  Deletion must be
 handled by the newly launched app version.

 (Turns out we never actually deleted the downloaded installer on Windows,
 filling the temporary directory.)

+++ install-mac.js
@@ -32,7 +32,7 @@ const fs = require('fs');
 const http = require('http');
 const { app, autoUpdater } = require('electron');
 
-async function install(updatePath, restart) {
+async function install(updatePath, restart, deleteAfterInstall) {
     console.log('Installing update');
 
     // We use the build-in updater for Mac installation for now, by creating a
@@ -102,7 +102,9 @@ async function install(updatePath, restart) {
             server.close();
             // Delete original update file.
             try {
-                fs.unlinkSync(updatePath);
+                if (deleteAfterInstall) {
+                    fs.unlinkSync(updatePath);
+                }
             } catch (err) {
                 console.error('Failed to delete update file');
                 // continue updating even if failed to delete
